chore(server): split and group file types into more logical units
